feat(gamification): remove weekly and monthly leaderboards
